+++ consensus/types/height_vote_set.go
@@ -1,4 +1,4 @@
-package consensus
+package types
 
 import (
 	"strings"
