+++ rpc/core/consensus.go
@@ -1,13 +1,10 @@
 package core
 
 import (
-	"encoding/json"
-
 	cm "github.com/tendermint/tendermint/consensus"
+	cstypes "github.com/tendermint/tendermint/consensus/types"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 	"github.com/tendermint/tendermint/types"
-
-	"github.com/pkg/errors"
 )
 
 // Get the validator set at the given block height.
@@ -85,22 +82,12 @@ func Validators(heightPtr *int) (*ctypes.ResultValidators, error) {
 // }
 // ```
 func DumpConsensusState() (*ctypes.ResultDumpConsensusState, error) {
-	roundStateBytes, err := json.Marshal(consensusState.GetRoundState())
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to marshal round state")
-	}
-
-	peerRoundStates := make(map[string]json.RawMessage)
-	for i, peer := range p2pSwitch.Peers().List() {
+	peerRoundStates := make(map[string]*cstypes.PeerRoundState)
+	for _, peer := range p2pSwitch.Peers().List() {
 		// TODO: clean this up?
 		peerState := peer.Get(types.PeerStateKey).(*cm.PeerState)
 		peerRoundState := peerState.GetRoundState()
-		peerRoundStateBytes, err := json.Marshal(peerRoundState)
-		if err != nil {
-			return nil, errors.Wrapf(err, "failed to marshal peer#%d round state", i)
-		}
-		peerRoundStates[peer.Key()] = json.RawMessage(peerRoundStateBytes)
+		peerRoundStates[peer.Key()] = peerRoundState
 	}
-
-	return &ctypes.ResultDumpConsensusState{json.RawMessage(roundStateBytes), peerRoundStates}, nil
+	return &ctypes.ResultDumpConsensusState{consensusState.GetRoundState(), peerRoundStates}, nil
 }
